requirements.txt
Django==2.0.8
cassandra-driver==3.13.0
djangorestframework==3.8.2
django-cassandra-engine==1.5.3
